refactor: consolidate all domain models into centralized models package
