+++ src/ade_compliance/models/axiom.py
@@ -1,45 +1,3 @@
-from datetime import datetime, timezone
-from enum import Enum
-from typing import Optional
+from . import Axiom, TraceLink, Violation, ViolationState
 
-from pydantic import BaseModel, Field
-
-
-class ViolationState(str, Enum):
-    NEW = "new"
-    ACKNOWLEDGED = "acknowledged"
-    RESOLVED = "resolved"
-    OVERRIDDEN = "overridden"
-
-
-class TraceLink(BaseModel):
-    source: str
-    target: str
-    type: str
-
-
-class Axiom(BaseModel):
-    id: str
-    name: str
-    category: str
-    severity: str
-    enabled: bool = True
-    description: Optional[str] = None
-
-
-class Violation(BaseModel):
-    axiom_id: str
-    file_path: str
-    message: str
-    severity: str = "medium"
-    timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc).replace(tzinfo=None))
-    state: ViolationState = ViolationState.NEW
-
-    def acknowledge(self):
-        self.state = ViolationState.ACKNOWLEDGED
-
-    def resolve(self):
-        self.state = ViolationState.RESOLVED
-
-    def override(self):
-        self.state = ViolationState.OVERRIDDEN
+__all__ = ["Axiom", "TraceLink", "Violation", "ViolationState"]
