+++ src/ade_compliance/models/decision.py
@@ -1,53 +1,3 @@
-import uuid
-from datetime import datetime, timezone
-from typing import List, Optional
+from . import Attestation, Decision, Override
 
-from pydantic import BaseModel, Field, model_validator
-
-
-class Decision(BaseModel):
-    axiom_id: str
-    rationale: str
-    criticality: str
-    timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc).replace(tzinfo=None))
-
-    @property
-    def requires_human_review(self) -> bool:
-        return self.criticality in ["high", "critical"]
-
-
-class Override(BaseModel):
-    id: str = Field(default_factory=lambda: str(uuid.uuid4()))
-    axiom_id: str
-    scope_type: str  # FILE | DIRECTORY | COMPONENT
-    scope_value: str
-    rationale: str
-    created_by: str
-    created_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc).replace(tzinfo=None))
-    expires_at: datetime
-    is_permanent: bool = False
-    permanent_justification: Optional[str] = None
-    revoked_at: Optional[datetime] = None
-
-    @model_validator(mode="after")
-    def validate_permanent(self) -> "Override":
-        if self.is_permanent and not self.permanent_justification:
-            raise ValueError("permanent_justification is required when is_permanent is True")
-        return self
-
-    @property
-    def is_active(self) -> bool:
-        if self.revoked_at is not None:
-            return False
-        if self.is_permanent:
-            return True
-        return datetime.now(timezone.utc).replace(tzinfo=None) < self.expires_at
-
-
-class Attestation(BaseModel):
-    agent_id: str
-    task_id: str
-    confidence: float
-    axioms_applied: List[str] = []
-    status: str = "pending"  # pending | passed | failed | escalated
-    timestamp: datetime = Field(default_factory=lambda: datetime.now(timezone.utc).replace(tzinfo=None))
+__all__ = ["Attestation", "Decision", "Override"]
